Funciona todo pero la estrategia de cargar y pintar la matriz desde flutter es demasiado ineficiente.

+++ lib/domain/matrix_repository.dart
@@ -5,7 +5,7 @@ class MatrixRepository {
 
   MatrixRepository({required this.apiService});
 
-  Future<List<List<int>>> getMatrix() async {
+  Future<List<List<double>>> getMatrix() async {
     return await apiService.fetchMatrix();
   }
 }
